create class for option report
